+++ engine/include/components/ParallaxComponent.hpp
@@ -14,25 +14,25 @@ namespace Engine {
 
     class ParallaxComponent : public Engine::Component {
     private:
-        Point<float> _start;
-        Point<float> _limit;
+        Point<double> _start;
+        Point<double> _limit;
 
     public:
         explicit ParallaxComponent() : _start({0, 0}), _limit({0, 0}), Engine::Component() {}
-        explicit ParallaxComponent(const Point<float> &start,const Point<float> &limit) : _start(start), _limit(limit),
+        explicit ParallaxComponent(const Point<double> &start,const Point<double> &limit) : _start(start), _limit(limit),
         Engine::Component() {}
 
-        [[nodiscard]] bool isLimitReached(const Point<float> &pos, const Vector<float> &direction) const {
+        [[nodiscard]] bool isLimitReached(const Point<double> &pos, const Vector<double> &direction) const {
             return (
                     (direction.x >= 0 && pos.x >= _limit.x) && (direction.y >= 0 && pos.y >= _limit.y) ||
                     (direction.x <= 0 && pos.x <= _limit.x) && (direction.y <= 0 && pos.y <= _limit.y)
             );
         }
 
-        [[nodiscard]] Point<float> getStart() const {return this->_start;}
-        [[nodiscard]] Point<float> getLimit() const {return this->_limit;}
-        void setStart(const Point<float> &start) {this->_start = start;}
-        void setLimit(const Point<float> &limit) {this->_limit = limit;}
+        [[nodiscard]] Point<double> getStart() const {return this->_start;}
+        [[nodiscard]] Point<double> getLimit() const {return this->_limit;}
+        void setStart(const Point<double> &start) {this->_start = start;}
+        void setLimit(const Point<double> &limit) {this->_limit = limit;}
     };
 
 }
