+++ server/game/components/ProjectileComponent.hpp
@@ -12,19 +12,19 @@ class ProjectileComponent : public Engine::Component
 {
 private:
     std::chrono::high_resolution_clock::time_point _birth;
-    float _damage;
-    float _lifetime;
+    double _damage;
+    double _lifetime;
 
 public:
     explicit ProjectileComponent() : _damage(0), _lifetime(0), Engine::Component() {}
-    explicit ProjectileComponent(float damage, float lifetime) : _damage(damage), _lifetime(lifetime),
+    explicit ProjectileComponent(double damage, double lifetime) : _damage(damage), _lifetime(lifetime),
     _birth(std::chrono::high_resolution_clock::now()), Engine::Component() {}
 
-    [[nodiscard]] float getDamage() const {return _damage;}
-    void setDamage(float damage) {_damage = damage;}
+    [[nodiscard]] double getDamage() const {return _damage;}
+    void setDamage(double damage) {_damage = damage;}
 
-    [[nodiscard]] float getLifetime() const {return _lifetime;}
-    void setLifetime(float lifetime) {_lifetime = lifetime;}
+    [[nodiscard]] double getLifetime() const {return _lifetime;}
+    void setLifetime(double lifetime) {_lifetime = lifetime;}
 
     [[nodiscard]] bool isLifetimeOver() const {return Engine::Timer::hasElapsed(_birth, _lifetime);}
 };
