+++ server/game/systems/EnemySystem.cpp
@@ -22,8 +22,8 @@ EnemySystem::EnemySystem(std::shared_ptr<Game> &game) : _game(game), Engine::Sys
 
 void EnemySystem::handleMovements(std::shared_ptr<Engine::Entity> &enemy)
 {
-    const float speed = 0.2;
-    Engine::Vector<float> dir = {0, 0};
+    const double speed = 0.2f;
+    Engine::Vector<double> dir = {0, 0};
 
     dir.x += 1;
     dir.y += 1;
