+++ server/game/systems/PlayerSystem.hpp
@@ -15,7 +15,7 @@ private:
     std::shared_ptr<Game> _game;
     std::shared_ptr<Engine::ATexture> _projectileTexture;
 
-    static void handleMoveAnimations(std::shared_ptr<Engine::Entity> &player, Engine::Vector<float> dir);
+    static void handleMoveAnimations(std::shared_ptr<Engine::Entity> &player, Engine::Vector<double> dir);
     static void handleMovements(std::shared_ptr<Engine::Entity> &player);
     void handleWeapon(std::shared_ptr<Engine::Entity> &player);
     static void handleCollisions(std::shared_ptr<Engine::Entity> &player);
